feat: componentize hero section and add image preview
