+++ apps/web/src/app/globals.css
@@ -75,7 +75,7 @@
 }
 
 @theme inline {
-  --font-safiro: var(--font-safiro);
+  --font-satoshi: var(--font-satoshi);
   --font-sans: var(--font-geist-sans);
   --font-mono: var(--font-geist-mono);
   --color-background: var(--background);
